Make babelrcRoots only match exact directories
Previously any path with a prefix of the directories specified would match, causing false positives, particularly in the case where the babelrcRoots included `"."`, since then all subdirectories are considered babelrc roots.

This introduces a new parameter, `matchPrefix`, to `patternToRegex` that specifies whether a prefix or exact match is required. This flag is then disabled when matching in `babelrcLoadEnabled`.

Fixes #12314.