Remove the cross-platform absolute path check
An absolute CommonJS module identifier is [defined to
be](https://nodejs.org/api/modules.html#modules_all_together) one that
starts with '/'. CommonJS module identifiers are platform independent.
That is, they are the same on every platform. Hence, there is no need
for a cross-platform absolute path check to determine if a CommonJS
module identifier represents an absolute module path or not.

This change also simplifies some of the regexes since the case of
starting with '/' is handled by the first conditional in
'standardizeName'.