Support merging sourcemaps when transformed file is fully replaced
A follow-up to #14246. I finally thought of a case where the
sourceFileName's content is fully replaced, leading to only the injected
file existing in the output sourcemap. This can happen where only one
output source is created, or multiple.

In the single output source case, we'd incorrectly associate the
mappings through the inputMap, even though none of the content actually
comes from there.

In the multiple source case, we'd silently fail and output empty mappings.

Both cases are now fixed, with the correct remapping being done through
in all possible output cases now.