Make sourcemap plainobject for use w / valueToNode
When the sourcemaps were switched to use "@ampproject/remapping", the type
of the inputSourceMap that gets passed around changed from being a plain js
object to a `class SourceMap` fromthe remapping package. This causes issues
with the @babel/types valueToNode function because that function is defined
to throw for objects that aren't plain-objects.

In practice I was seeing this error after upgrading babel when running code-
coverage, and I saw a similar error reported here: https://github.com/vuejs/vue-jest/issues/450 before
deciding to try to track the error down myself.