Support `indices.groups` when compiling named groups in regexps (#15092)
Add support for indices.groups

If a RegExp has the /d flag (hasIndices), match objects should have an
extra property .indices with an array of the character indices that each
subgroup matched. The indices array itself has a property .groups giving
the matched character indices (or undefined) for each named capturing
group.

This doesn't do anything to enable support for the /d flag, the
hasIndices property, or the returned indices object, if the environment
does not support it, but it will populate .indices.groups in _wrapRegExp
if .indices is present, so that the
transform-named-capturing-groups-regex and
proposal-duplicate-named-capturing-groups-regex plugins can take
advantage of that support if it exists.