feature: new assumption privateFieldsAsSymbols, fixes #15389
When babel-helpers are inlined, classPrivateFieldLooseKey()
defines a global variable `id` and initialized it to zero
in every generated file. This can lead to name clashes
between #-private members in subclass and superclass,
which again leads to incorrect semantics.
In #15389, it has been discussed to fix this, but
eventually the suggestion was to introduce a new assumption
`privateFieldsAsSymbols` and deprecate (and later remove)
the buggy `privateFieldsAsProperties`.
In the new solution, instead of trying to produce a
globally unique private property name string, a `Symbol`
is created. To use the new assumption, in environments
where `Symbol` is not available (Internet Explorer), a
polyfill must be included.