Introduce `bugfix-safari-class-field-initializer-scope`
This is a workaround for https://bugs.webkit.org/show_bug.cgi?id=236843, and aims to resolve #14289.

It adds the new compat-table check to the requirements for transform-class-properties, which means that transform will now apply for Safari < 16 (previously Safari < 14.5).

It also introduces a bugfix transformation which can be used as an alternative to transform-class-properties on Safari 15. This bugfix transformations wraps potentially-affected initializers in an IIFE. A number of shortcuts are added to avoid transforming initializers which are guaranteed to be unaffected by the bug.