fix: improve ts-only declaration parsing (#17491)
* fix: improve ts-only declaration parsing

Handle TS declaration parsing in parseStatementContent such that we can align to the tsc behaviour when parsing `type as =` or `module as {}`. In both cases tsc treats `as` as an identifier name, but previously Babel parser treated it as rational operator because the ts-only declaration parsing was delayed after a JS expression is formed.

* remove unused directive

* update Babel 8 test fixtures

* address review comments

* refactor: classify tsParseDeclaration by type